pyproject.toml
[tool.poetry]
name = "pipelines"
version = "0.1.0"
description = ""
authors = ["Leonid Shvechikov <shvechikov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1.3"
pandas = "^1.5.3"
psycopg2-binary = "^2.9.5"
pytest = "^7.2.2"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
